% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acad_year}
\alias{acad_year}
\title{Get the academic year of a term}
\usage{
acad_year(x)
}
\arguments{
\item{x}{An object of class 'term'}
}
\value{
Character
}
\description{
Get the academic year of a term
}
