% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{as_term}
\alias{as_term}
\title{Cast an object to term}
\usage{
as_term(x, origin = c("sims", "cs"))
}
\arguments{
\item{x}{An object}

\item{origin}{Either 'cs' or 'sims'}
}
\value{
An object of class 'term'
}
\description{
Cast an object to term
}
