% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_next}
\alias{get_next}
\alias{get_last}
\title{Get the next or last n terms}
\usage{
get_next(x, n, keep = NA)

get_last(x, n, keep = NA)
}
\arguments{
\item{x}{an object of class 'term'}

\item{n}{how many terms}

\item{keep}{which terms to keep (ending character)}
}
\value{
term sequence
}
\description{
Get the next or last n terms
}
\examples{
x <- term(20164)
get_next(x, 5)
get_next(x, 10, keep = "4")
get_last(x, 5)
}
