% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_term}
\alias{label_term}
\title{Label a term object}
\usage{
label_term(x)
}
\arguments{
\item{x}{An object of class 'term'}
}
\value{
Character, e.g. "Fall 2018"
}
\description{
Label a term object
}
\examples{
label_term(term(20194))
}
