% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{mathematics}
\alias{mathematics}
\alias{median.term}
\alias{min.term}
\alias{max.term}
\title{Math stuff for term objects}
\usage{
\method{median}{term}(x, ...)

\method{min}{term}(x, na.rm = FALSE, ...)

\method{max}{term}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An object of class 'term'}

\item{...}{Additional args passed to various functions}

\item{na.rm}{Logical. Should missing values be removed?}
}
\description{
Math stuff for term objects
}
