% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{new_term}
\alias{new_term}
\title{Internal constructor to create term type
Can be cast from either character or double}
\usage{
new_term(x = double(), origin = c("sims", "cs"))
}
\description{
Internal constructor to create term type
Can be cast from either character or double
}
\keyword{internal}
