% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+F\%}
\alias{\%+F\%}
\alias{\%-F\%}
\alias{\%+S\%}
\alias{\%-S\%}
\title{Get next fall/spring sequences}
\usage{
lhs \%+F\% rhs

lhs \%-F\% rhs

lhs \%+S\% rhs

lhs \%-S\% rhs
}
\arguments{
\item{lhs}{an object of class 'term'}

\item{rhs}{an integer}
}
\value{
sequence of terms
}
\description{
Get next fall/spring sequences
}
