% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{term}
\alias{term}
\title{A class for SIS terms}
\usage{
term(x, origin = c("sims", "cs"))
}
\arguments{
\item{x}{A character or double vector}

\item{origin}{either 'cs' or 'sims'}
}
\description{
A class for SIS terms
}
