% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{term_duration}
\alias{term_duration}
\title{Get the number of terms in between two terms}
\usage{
term_duration(x, y)
}
\arguments{
\item{x}{a term object}

\item{y}{a term object}
}
\value{
integer
}
\description{
Get the number of terms in between two terms
}
\examples{
x <- term(20104)
y <- term(20144)
term_duration(x, y)
term_duration(y, x)
}
