% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R, R/coercion.R, R/comparison.R,
%   R/termtools-package.R
\name{vec_cast.term}
\alias{vec_cast.term}
\alias{vec_ptype2.term}
\alias{vec_proxy_compare.term}
\alias{vec_proxy_equal.term}
\alias{vctrs-compat}
\title{vctrs compatibility functions for math, casting, coercion, and comparison}
\usage{
\method{vec_cast}{term}(x, to, ...)

\method{vec_ptype2}{term}(x, y, ...)

\method{vec_proxy_compare}{term}(x, ...)

\method{vec_proxy_equal}{term}(x, ...)
}
\arguments{
\item{x, y}{Vectors.}

\item{to}{Type to cast to.}

\item{...}{Arguments passed on to further methods.}

\item{op}{Arithmetic operation.}

\item{.fn}{Mathematical function.}

\item{.x}{Vector.}
}
\description{
These functions are the extensions that allow termtools vectors to
work with vctrs.
}
